+++ test-data/yaml/validation/test_wf_00.gxwf.yml
@@ -1,5 +1,2 @@
 class: GalaxyWorkflow
-label: Test Workflow Without Steps
-inputs: {}
-outputs: {}
-steps: {}
+label: Test workflow with missing required fields
